Created new files for FAQ section, modified the logic of the handler
