+++ src/bot/constants/text.py
@@ -1,10 +1,3 @@
-from bot.constants.links import (
-    STORAGE_LINK,
-    STUDIES_LINK,
-    VOLUNTARISM_LINK,
-)
-
-
 STOP_MESSAGE = "Работа приложения остановлена."
 HELP_MESSAGE = "Написать в поддержку"
 START_MESSAGE = (
@@ -26,110 +19,3 @@ PASSED_THE_TEST = (
 STICKER_ID: str = (
     "CAACAgIAAxkBAAIDwGUEItYpuU5DXDRigDA1M9SL-AdcAALTJAACHRAwSj-yFNWnj6hDMAQ"
 )
-BACK_TO_MENU = (
-    "Выбери интересующую тебя тему из списка "
-    "в меню и я помогу тебе найти нужную информацию."
-)
-FAQ_MESSAGE = "FAQ. Выберите интересующий вопрос"
-ORG_QSTS_MESSAGE = f"""
-*У кого я могу узнать график работы?*
-График работы сообщается тебе в оффере. Для большинства офисных работников
-в трудовом договоре прописан график - 5/2 с возможностью работы из дома
-1 день в неделю. По согласованию с руководителем количество дней на удаленке
-может быть увеличено.
-
-В Фонде нет фиксированных часов работы, но мы стараемся быть на связи
-с 10 до 19. Мастерская Фонда работает с 12 до 21: Центр сортировки в Москве
-с 9 до 17,Центр сортировки и переработки в Костроме с 08:30 до 17:30.
-
-*Как и когда выдают зарплату?*
-Зарплата начисляется на вашу личную банковскую карту 2 раза в месяц:
-20-го числа аванс, 5-го — остальная часть зарплаты.
-
-*Что делать, если заболел(а)?*
-Сперва предупреди своего руководителя (чем раньше, тем лучше).
-Если есть необходимость уйти на больничный, то нужно вызвать или посетить
-врача. После выздоровления, отправь электронный больничный лист в бухгалтерию.
-Если ты заболел(а) во время отпуска, то вы можно либо продлить свой отдых на
-соответствующее количество дней, либо перенести неиспользованную часть
-отпуска на другое время. О болезни и своем решении продлить или перенести
-отпуск сообщи непосредственному руководителю.
-
-*Есть ли в Фонде ДМС?*
-Да,  добровольное медицинское страхование предоставляется всем
-сотрудникам, для которых Фонд является основным местом работы.
-ДМС подключается после года работы в Фонде и включает в себя стационарную
-помощь и помощь на дому (в зависимости от программы). Уведомление о
-готовности ДМС отправляется автоматически на рабочую почту с указанием
-программы и списка доступных клиник.
-"""
-VOLUNTARISM_MESSAGE = (
-    f"Все о волонтёрстве в Фонде читай "
-    f"[в FAQ по волонтерству]({VOLUNTARISM_LINK})"
-)
-STUDIES_MESSAGE = (
-    f"Все об обучении и развитии сотрудников читай "
-    f"[в гайде по обучению]({STUDIES_LINK})"
-)
-VACATION_MESSAGE = f"""
-Ежегодный оплачиваемый отпуск предоставляется сотруднику по истечении
-6 месяцев непрерывной работы и составляет 28 календарных дней.
-
-Даты отпуска согласовываются с непосредственным руководителем на год
-и фиксируются в графике отпусков. Дату отпуска можно перенести,
-если у тебя изменились обстоятельства, в таком случае,
-также подходишь к руководителю.
-Согласованные даты или изменения обязательно передаем HR.
-За неделю до отпуска необходимо написать заявление, чтобы отпускные
-пришли вовремя (приходят они за 3 дня до даты отпуска).
-"""
-WORK_PROC_MESSAGE = f"""
-*Какой у Фонда юридический адрес и почтовый адрес?*
-Юридический адрес: 129085, Москва г, ул. Годовикова, д. 6,
-кв. этаж 1 помещение IIА, VII, комнаты 1-7, 7 А, 8, 9/1-6, 6А, 7
-
-*Как найти склад Фонда?*
-[Вот подробная схема.]({STORAGE_LINK})
-
-*Есть ли дресс-код в фонде?*
-Мы уважаем твою индивидуальность и право на свободу самовыражения,
-поэтому дресс-кода у нас нет.
-
-
-*Как организовать отправку и доставку корреспонденции в регион?*
-Процесс обмена документами с регионами обычно осуществляется либо по почте,
-либо через московский центр сортировки. Передай документы нашему
-администратору офиса с четкими инструкциями,
-и документы попадут в нужное место.
-"""
-ADMIN_QSTS_MESSAGE = f"""
-*Как заказать канцелярские товары?*
-Обратись в административный отдел Фонда. если тебе что-то нужно.
-
-*Что делать в случае неисправности оргтехники?*
-Оргтехникой тоже заведует административный отдел.
-
-*Надо ли приносить с собой свою посуду, если я обедаю на общей кухне?*
-На кухне есть необходимая посуда, но для своего комфорта можешь взять свою.
-Если вы большой любитель чая и кофе, личная кружка будет не лишней.
-"""
-DOCFLOW_MESSAGE = f"""
-*У кого можно заказать справку 2-НДФЛ?*
-Оформлением подобных справок занимается бухгалтерия. Справки с места работы
-или заверенную копию трудовой книжки можно получить у HR.
-
-*Как и где получить шаблоны необходимых документов?*
-Проверь раздел "Регламенты и документы", если там нет нужного тебе
-регламента, обратись с вопросом к HR.
-"""
-BUSINESS_TRIP_MESSAGE = f"""
-*Как забронировать гостиницу и трансфер для командировки?*
-Сотрудники, которые едут в командировку, могут бронировать перелёт или
-переезд и проживание через административный отдел или самостоятельно
-с последующей компенсацией.
-
-*колько мне положено командировочных на поездку?*
-За каждый день командировки тебе выдается 700 рублей суточных.
-Чтобы они были начислены вовремя, необходимо оформлять командировку
-хотя бы за 3 дня до поездки.
-"""
