+++ src/bot/handlers.py
@@ -5,24 +5,18 @@ from telegram.ext import (
     filters,
 )
 
-from bot.constants.state import CHECK, FAQ, MENU
+from bot.constants.state import (
+    CHECK,
+    FAQ,
+    MAIN_MENU,
+)
 from bot.conversations.main_application import (
     greeting_callback,
     done_callback,
     check_the_secret_word_callback,
-    faq_callback,
-    back_button_callback,
+    main_menu_actions_callback,
 )
-from bot.conversations.menu_application import (
-    org_qsts_callback,
-    voluntarism_callback,
-    studies_callback,
-    vacation_callback,
-    work_proc_callback,
-    admin_qsts_callback,
-    docflow_callback,
-    business_trip_callback,
-)
+from bot.conversations.faq_application import faq_callback
 
 
 conv_handler = ConversationHandler(
@@ -31,31 +25,11 @@ conv_handler = ConversationHandler(
         CHECK: [
             MessageHandler(filters.TEXT, check_the_secret_word_callback),
         ],
-        MENU: [
-            MessageHandler(filters.Text(["FAQ"]), faq_callback),
+        MAIN_MENU: [
+            MessageHandler(filters.Text(["FAQ"]), main_menu_actions_callback),
         ],
         FAQ: [
-            MessageHandler(
-                filters.Text(["Организационные вопросы"]), org_qsts_callback
-            ),
-            MessageHandler(
-                filters.Text(["Волонтёрство"]), voluntarism_callback
-            ),
-            MessageHandler(filters.Text(["Обучение"]), studies_callback),
-            MessageHandler(filters.Text(["Отпуск"]), vacation_callback),
-            MessageHandler(
-                filters.Text(["Рабочий процесс"]), work_proc_callback
-            ),
-            MessageHandler(
-                filters.Text(["Административные вопросы"]), admin_qsts_callback
-            ),
-            MessageHandler(
-                filters.Text(["Оформление документов"]), docflow_callback
-            ),
-            MessageHandler(
-                filters.Text(["Командировки"]), business_trip_callback
-            ),
-            MessageHandler(filters.Text(["Назад"]), back_button_callback),
+            MessageHandler(filters.Text(), faq_callback),
         ],
     },
     fallbacks=[MessageHandler(filters.Regex("^Done$"), done_callback)],
